// frontend-app/src/assets/styles/tailwind/plugins/components/progress/progress.types.ts
import type { PropertyVariant } from '../../../types'

export interface ProgressProps extends Record<string, unknown> {
  color?:
    | 'primary'
    | 'info'
    | 'success'
    | 'warning'
    | 'danger'
    | 'light'
    | 'dark'
    | 'black'
  contrast?: 'default' | 'contrast'
  rounded?: 'none' | 'sm' | 'md' | 'lg' | 'full'
  size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl'
  value?: number
  max?: number
  classes?: {
    wrapper?: string | string[]
    progress?: string | string[]
  }
}

export interface ProgressEvents {}

export interface ProgressSlots {}

export type ProgressAttrs = ProgressProps & ProgressEvents & ProgressSlots
export type ProgressVariant<T> = PropertyVariant<T, ProgressProps>